Implement a Root-Level Request Logger Middleware
Build a simple logger. For every request, it should log in the console a string taking the following format: method path - ip
Hint: Express evaluates functions in the order they appear in the code. This is true for middleware too. If you want it to work for all the routes, it should be mounted before the